create manager account

+++ src/main/java/nl/fontys/s3/carenestproject/domain/classes/Role.java
@@ -10,5 +10,5 @@ public enum Role {
     PATIENT(2),
     CARETAKER(3);
 
-    private final int value;
+    private final long value;
 }
